release 20171222

+++ make_MagOS/files/patches/rootfs/MagOS/usr/lib/magos/rootfs-patches/MagOS/MagOS-kde.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+[ ! -f /usr/bin/startkde -o -d /etc/xdg/plasma-workspace ] && exit 0
 ln -sf /usr/share/magos/kde4/kde4rc                        /etc/alternatives/kde4-config
 ln -sf /usr/share/magos/kde4/share/config/kdm/backgroundrc /etc/alternatives/kdm4-background-config
 ln -sf /usr/share/magos/kde4/share/config/kdm/kdmrc        /etc/alternatives/kdm4-config
@@ -15,15 +16,5 @@ if [ -d /usr/share/apps/desktoptheme ] ;then
   fi
  done
 fi
-
-#for Xsession based dm's
-PFP=/etc/X11/wmsession.d/01KDE4
-cat >>$PFP <<EOF
-NAME=KDE4
-DESC=KDE Desktop
-EXEC=/usr/bin/startkde
-SCRIPT:
-exec /usr/bin/startkde
-EOF
 
 exit 0
